# specs/002-rag-chatbot-migration/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: RAG Chatbot API
  description: API for the RAG Chatbot system with OpenAI Agent SDK, Qdrant, and Cohere embeddings
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://your-backend-api.com
    description: Production server

paths:
  /embed:
    post:
      summary: Create embeddings and upsert to Qdrant
      description: Process markdown files, generate embeddings using Cohere, and store in Qdrant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                source_path:
                  type: string
                  description: Path to markdown files to process
                  example: "/docs/textbook"
                force_recreate:
                  type: boolean
                  description: Whether to recreate embeddings even if they exist
                  default: false
              required:
                - source_path
      responses:
        '200':
          description: Embedding process completed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  chunks_processed:
                    type: integer
                    example: 150
                  collection_name:
                    type: string
                    example: "book_chunks"
        '400':
          description: Bad request - invalid parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error during embedding process
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /rag:
    post:
      summary: RAG query with context retrieval
      description: Query the system using RAG to retrieve relevant chunks and generate response
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                query:
                  type: string
                  description: The question or query to answer
                  example: "What is the main concept of humanoid robotics?"
                context:
                  type: string
                  description: Optional additional context for the query
                max_chunks:
                  type: integer
                  description: Maximum number of chunks to retrieve
                  default: 5
                  minimum: 1
                  maximum: 10
              required:
                - query
      responses:
        '200':
          description: RAG query processed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  answer:
                    type: string
                    description: The generated answer
                    example: "Humanoid robotics focuses on creating robots with human-like characteristics..."
                  sources:
                    type: array
                    items:
                      type: string
                    description: IDs of chunks used as sources
                    example: ["chunk_1", "chunk_2", "chunk_3"]
                  confidence:
                    type: number
                    format: float
                    description: Confidence score of the response
                    minimum: 0.0
                    maximum: 1.0
                    example: 0.85
        '400':
          description: Bad request - invalid query
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Embeddings not found - run embed pipeline first
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error during RAG process
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /rag/selected:
    post:
      summary: Selected text RAG query
      description: Answer question based only on provided selected text, bypassing Qdrant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                selected_text:
                  type: string
                  description: The text selected by the user
                  example: "The main principle of humanoid robotics is to create robots with human-like characteristics and movements."
                query:
                  type: string
                  description: Question about the selected text
                  example: "What is the main principle mentioned in this text?"
              required:
                - selected_text
                - query
      responses:
        '200':
          description: Selected text RAG processed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  answer:
                    type: string
                    description: The generated answer based only on selected text
                    example: "The main principle is to create robots with human-like characteristics and movements."
                  sources:
                    type: array
                    items:
                      type: string
                    description: Empty array since no external sources were used
                    example: []
        '400':
          description: Bad request - invalid parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error during selected text RAG process
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/conversation:
    post:
      summary: Chat conversation endpoint
      description: Handle chat conversation with the AI agent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                message:
                  type: string
                  description: The user's message
                  example: "Can you explain the concept of humanoid robotics?"
                session_id:
                  type: string
                  description: Optional session identifier
                context:
                  type: object
                  description: Additional context for the conversation
              required:
                - message
      responses:
        '200':
          description: Chat message processed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  response:
                    type: string
                    description: The agent's response
                    example: "Humanoid robotics is a branch of robotics focused on creating robots with human-like characteristics..."
                  session_id:
                    type: string
                    description: Session identifier (newly created if not provided)
                    example: "sess_abc123"
                  sources:
                    type: array
                    items:
                      type: string
                    description: Sources used in the response
                    example: ["chunk_1", "chunk_2"]
        '400':
          description: Bad request - invalid message
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error during conversation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /health:
    get:
      summary: Health check endpoint
      description: Check if the service is running and healthy
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "healthy"
                  timestamp:
                    type: string
                    format: date-time
                    example: "2025-12-10T10:00:00Z"
        '503':
          description: Service is unhealthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
          example: "Embedding not found: Re-run embed pipeline"
        code:
          type: string
          description: Error code
          example: "EMBEDDING_NOT_FOUND"
        details:
          type: object
          description: Additional error details
          example: {}